+++ src/operations/install.rs
@@ -7,19 +7,31 @@ use std::os::unix::fs::symlink;
 use std::path::{Path, PathBuf};
 use url::Url;
 
-fn install_git_url(data_item_path: &Path, item_name: &str, item_git_url: &str) -> Result<()> {
+fn install_git_url(
+    data_item_path: &Path,
+    item_name: &str,
+    item_git_url: &str,
+    is_quiet: bool,
+) -> Result<()> {
     if !data_item_path.is_dir() {
         git_clone(item_git_url, data_item_path)?;
 
-        println!("{} installed", item_name);
-    } else {
+        if !is_quiet {
+            println!("{} installed", item_name);
+        }
+    } else if !is_quiet {
         println!("{} already installed", item_name);
     }
 
     Ok(())
 }
 
-fn install_dir(data_item_path: &Path, item_name: &str, item_path: &Path) -> Result<()> {
+fn install_dir(
+    data_item_path: &Path,
+    item_name: &str,
+    item_path: &Path,
+    is_quiet: bool,
+) -> Result<()> {
     if item_path.exists() && !item_path.is_dir() {
         return Err(anyhow!(
             "{} is not a symlink to a directory. Please remove it and try again",
@@ -35,17 +47,22 @@ fn install_dir(data_item_path: &Path, item_name: &str, item_path: &Path) -> Resu
                         return Err(anyhow!("Error trying to remove symlink at \"{}\". Remove it manually and try again", data_item_path.display()));
                     }
                     symlink(item_path, data_item_path)?;
-                    println!("{} already installed", item_name);
+
+                    if !is_quiet {
+                        println!("{} already installed", item_name);
+                    }
                 }
             }
             Err(_) => {
-                println!("err");
                 return Err(anyhow!("\"{}\" is a not a symlink, but according to your config it should be. Please remove this directory and try again", data_item_path.display()));
             }
         }
     } else {
         symlink(item_path, data_item_path)?;
-        println!("{} installed", item_name);
+
+        if !is_quiet {
+            println!("{} installed", item_name);
+        }
     }
 
     Ok(())
@@ -55,8 +72,7 @@ fn install_dir(data_item_path: &Path, item_name: &str, item_path: &Path) -> Resu
 ///
 /// Clones the provided config repositories and ensures everything is ready for when the user runs
 /// any other command
-pub fn install(config_path: &Path, data_path: &Path) -> Result<()> {
-    // Hooks
+pub fn install(config_path: &Path, data_path: &Path, is_quiet: bool) -> Result<()> {
     let config = Config::read(config_path)?;
     let items = config.items.unwrap_or_default();
     let hooks_path = data_path.join(REPO_DIR);
@@ -66,14 +82,24 @@ pub fn install(config_path: &Path, data_path: &Path) -> Result<()> {
         let item_path = PathBuf::from(item.path.as_str());
 
         match Url::parse(item.path.as_str()) {
-            Ok(_) => install_git_url(&data_item_path, item.name.as_str(), item.path.as_str())?,
-            Err(_) => install_dir(&data_item_path, item.name.as_str(), &item_path)?,
+            Ok(_) => install_git_url(
+                &data_item_path,
+                item.name.as_str(),
+                item.path.as_str(),
+                is_quiet,
+            )?,
+            Err(_) => install_dir(&data_item_path, item.name.as_str(), &item_path, is_quiet)?,
         }
     }
 
-    // Schemes
     let schemes_repo_path = hooks_path.join(SCHEMES_REPO_NAME);
-    install_git_url(&schemes_repo_path, SCHEMES_REPO_NAME, SCHEMES_REPO_URL)?;
+
+    install_git_url(
+        &schemes_repo_path,
+        SCHEMES_REPO_NAME,
+        SCHEMES_REPO_URL,
+        is_quiet,
+    )?;
 
     Ok(())
 }